chore: add comments across source
chore: refactor source code
