+++ src/utils.ts
@@ -5,7 +5,8 @@ import * as mod from "https://deno.land/std@0.154.0/log/mod.ts";
 export const logger = await getLogger();
 
 /**
- * @returns
+ * Setup logger
+ * @returns mod.Logger
  */
 export async function getLogger(): Promise<mod.Logger> {
   await mod.setup({
@@ -23,98 +24,4 @@ export async function getLogger(): Promise<mod.Logger> {
   });
 
   return mod.getLogger();
-}
-
-// File System
-
-/**
- * @param filePath
- * @returns
- */
-export function listDirectories(filePath: string): string[] {
-  const directoryNames: string[] = [];
-
-  for (const _directory of Deno.readDirSync(filePath)) {
-    if (_directory.isDirectory) {
-      directoryNames.push(_directory.name);
-    }
-  }
-
-  if (directoryNames.length === 0) {
-    logger.warning(`Application terminated`);
-    throw new Error(`No directories found in ${filePath}`);
-  }
-
-  return directoryNames;
-}
-
-/**
- * @param filePath
- * @returns
- */
-export function listFiles(filePath: string): string[] {
-  const fileNames: string[] = [];
-
-  try {
-    for (const _file of Deno.readDirSync(filePath)) {
-      if (_file.isFile) {
-        fileNames.push(_file.name);
-      }
-    }
-  } catch (err) {
-    logger.warning("Application terminated");
-    throw err;
-  }
-
-  if (fileNames.length === 0) {
-    logger.warning(`Application terminated`);
-    throw new Error(`No files found in ${filePath}`);
-  }
-
-  return fileNames;
-}
-
-/**
- * @param filePath
- * @returns
- */
-export function getFileContent(filePath: string): string {
-  const decoder = new TextDecoder("utf-8");
-
-  try {
-    return decoder.decode(Deno.readFileSync(filePath));
-  } catch (err) {
-    logger.warning(`Application terminated`);
-    throw err;
-  }
-}
-
-/**
- * @param filePath
- */
-export function testFilePath(filePath: string): void {
-  try {
-    Deno.statSync(filePath);
-  } catch (err) {
-    logger.warning(`Application terminated`);
-    throw err;
-  }
-}
-
-/**
- * @param filePath
- * @param content
- */
-export function writeJsonFile(filePath: string, content: any): void {
-  const encoder = new TextEncoder();
-
-  try {
-    Deno.writeFileSync(
-      filePath,
-      encoder.encode(JSON.stringify(content, null, 2)),
-    );
-  } catch (err) {
-    logger.warning(`Application terminated`);
-    throw err;
-  }
 }
